+++ __tests__/focus-prism.test.tsx
@@ -1,7 +1,9 @@
 import React, { StrictMode } from 'react'
 import { fireEvent, render } from '@testing-library/react'
-import { useAtom } from 'jotai/react'
+import { expectTypeOf } from 'expect-type'
 import { atom } from 'jotai/vanilla'
+import type { SetStateAction, WritableAtom } from 'jotai/vanilla'
+import { useAtom } from 'jotai/react'
 import * as O from 'optics-ts'
 import { focusAtom } from '../src/index'
 
@@ -65,4 +67,54 @@ it('atoms that focus on no values are not updated', async () => {
   fireEvent.click(getByText('button'))
   await findByText('count:')
   await findByText('bigAtom: {}')
+})
+
+type BillingData = {
+  id: string
+}
+
+type CustomerData = {
+  id: string
+  billing: BillingData[]
+  someDynamicData?: string
+}
+
+it('typescript should work well with nested arrays containing optional values', async () => {
+  const customerListAtom = atom<CustomerData[]>([])
+
+  const foundCustomerAtom = focusAtom(customerListAtom, (optic) =>
+    optic.find((el) => el.id === 'some-invalid-id')
+  )
+
+  const derivedAtom = focusAtom(foundCustomerAtom, (optic) => {
+    const result = optic
+      .valueOr({ billing: [] } as unknown as CustomerData)
+      .prop('billing')
+      .find((el) => el.id === 'some-invalid-id')
+
+    return result
+  })
+
+  expectTypeOf(derivedAtom).toMatchTypeOf<
+    WritableAtom<BillingData | undefined, [SetStateAction<BillingData>], void>
+  >()
+})
+
+it('should work with promise based atoms with "undefined" value', async () => {
+  const customerBaseAtom = atom<CustomerData | undefined>(undefined)
+
+  const asyncCustomerDataAtom = atom(
+    async (get) => get(customerBaseAtom),
+    (_, set, nextValue: CustomerData) => {
+      set(customerBaseAtom, nextValue)
+    }
+  )
+
+  const focusedPromiseAtom = focusAtom(asyncCustomerDataAtom, (optic) =>
+    optic.optional()
+  )
+
+  expectTypeOf(focusedPromiseAtom).toMatchTypeOf<
+    WritableAtom<CustomerData | undefined, [SetStateAction<CustomerData>], void>
+  >()
 })