+++ __tests__/focus-traversal.test.tsx
@@ -1,7 +1,9 @@
 import React, { StrictMode } from 'react'
 import { fireEvent, render } from '@testing-library/react'
+import { expectTypeOf } from 'expect-type'
 import { useAtom } from 'jotai/react'
 import { atom } from 'jotai/vanilla'
+import type { SetStateAction, WritableAtom } from 'jotai/vanilla'
 import * as O from 'optics-ts'
 import { focusAtom } from '../src/index'
 
@@ -34,4 +36,51 @@ it('updates traversals', async () => {
   fireEvent.click(getByText('button'))
   await findByText('count: 6,7')
   await findByText('bigAtom: [{"a":6},{},{"a":7}]')
+})
+
+type BillingData = {
+  id: string
+}
+
+type CustomerData = {
+  id: string
+  billing: BillingData[]
+  someOtherData: string
+}
+
+it('typescript should accept "undefined" as valid value for traversals', async () => {
+  const customerListListAtom = atom<CustomerData[][]>([])
+
+  const nonEmptyCustomerListAtom = focusAtom(customerListListAtom, (optic) =>
+    optic.find((el) => el.length > 0)
+  )
+
+  const focusedPromiseAtom = focusAtom(nonEmptyCustomerListAtom, (optic) => {
+    const result = optic.valueOr([]).elems()
+    return result
+  })
+
+  expectTypeOf(focusedPromiseAtom).toMatchTypeOf<
+    WritableAtom<CustomerData[], [SetStateAction<CustomerData>], void>
+  >()
+})
+
+it('should work with promise based atoms with "undefined" value', async () => {
+  const customerBaseAtom = atom<CustomerData[] | undefined>(undefined)
+
+  const asyncCustomerDataAtom = atom(
+    async (get) => get(customerBaseAtom),
+    (_, set, nextValue: CustomerData[]) => {
+      set(customerBaseAtom, nextValue)
+    }
+  )
+
+  const focusedPromiseAtom = focusAtom(asyncCustomerDataAtom, (optic) => {
+    const result = optic.valueOr([]).elems()
+    return result
+  })
+
+  expectTypeOf(focusedPromiseAtom).toMatchTypeOf<
+    WritableAtom<CustomerData[], [SetStateAction<CustomerData>], void>
+  >()
 })