+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [0.2.1] - 2023-01-29
+### Changed
+- refactor: simplify memoize
+- fix: accept 'undefined' as valid value for Prisms, Lens, ISO and Equivalence #5
+
 ## [0.2.0] - 2022-11-15
 ### Added
 - Move back from the official Jotai library